requirements.txt
Jinja2
Markdown
MarkupSafe
PyMuPDF
PyYAML
Pygments
SQLAlchemy
Werkzeug
absl-py
accelerate
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
astunparse
attrs
beautifulsoup4
bs4
certifi
cffi
charset-normalizer
click
contourpy
cycler
dataclasses-json
dirtyjson
diskcache
distro
fastapi
filelock
flatbuffers
fonttools
frozenlist
fsspec
gast
google-pasta
greenlet
grpcio
guidance
h11
h5py
html2text
httpcore
httpx
huggingface-hub
idna
interegular
jax
jaxlib
jiter
joblib
jsonpatch
jsonpointer
keras
kiwisolver
langchain
langchain-core
langchain-text-splitters
langsmith
libclang
lm-format-enforcer
markdown-it-py
marshmallow
matplotlib
mdurl
mediapipe
minijinja
ml-dtypes
mpmath
multidict
mypy-extensions
namex
nest-asyncio
networkx
nltk
numpy
openai
opencv-contrib-python
opencv-python
opt-einsum
optree
ordered-set
orjson
packaging
pandas
pillow
platformdirs
protobuf
psutil
pycparser
pydantic
pydantic_core
pydot
pyformlang
pyparsing
pypdf
python-dateutil
python-dotenv
pytz
regex
requests
rich
safetensors
scikit-learn
scipy
sentence-transformers
six
sniffio
sounddevice
soupsieve
starlette
striprtf
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
termcolor
threadpoolctl
tiktoken
tokenizers
torch
tqdm
transformers
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
wheel
wrapt
yarl